Fix for Navigation via arrow keys
Fix for Navigation of the GUI via arrow keys.  Currently a number of arrow key moves with crash the GUI due to references which have since been deleted from the build or do not get mapped on Raspberry.  Listing fixes all crashes and a couple of missing entries.  Still working on why two of the RAM options (Z3 fast and RTG) are not selectable via arrow keys.