fix compilation without gperftools (#87)
* fix compilation without gperftools

the -ltcmalloc get pulled in by gperftools

* fix buffer overflow

This might not be the last one...

* fix potential buffer overflow

Some still remain because I'm unsure of their function.
Just simple fixes for now.

* fix even more potential overflows

Those need testing. It seems a lot variables are hardcoded there
but since it matches NAME_MAX it should be safe to use 255 as a
value.