Amiberry GUI on Android comes up (#563)
* Uses C++14 standard lib instead of GNU

GNU lead to compilation errors.

* Removes variables

Those are in the main Android.mk now and relative.

* Creates header file for amiberry_filesys for reuse in other files

* Adds prefix_with_application_directory_path for Android support

* Uses path prefixer to fix some crashes on Android

* Adds pattern for Android Studio (jetbrains) working file

* Removes unused compiler directive LIBMPEG2_PATH from Makefile

* Fixes startup behavior for Android

We are using the applications directory as our working directory now,
because SDCARD access will be limited in the future of Android.
Coming with that, we're also relying on `start_path_data` being
set correctly at startup instead of working around it during
loading of settings.

* Workaround for crash during startup

* Workaround for another crash during startup on Android

* Adds new code file scp.cpp to Android build

* Adds orig files to gitignore

* Fixes behaviour of prefix_with_application_directory_path

* Enables redirection of logs to Android's logcat

* Fixes misspelled compiler directive