Only react to input events we care about
It seems that at least under X11, SDL2 sends events to the window constantly even when no user input has occurred.
This resulted in 100% cpu usage when having the GUI open, since for every event detected we were redrawing the screen for no reason.
With this change, we monitor only the kind of events we want to react upon, and perform the screen update when one of those is triggered.