Implemented option to map Autofire button in Custom Controls
You can now map a button on your controller to be the main Fire, using Autofire mode. This can be separate from the normal Fire button you already have, and which can be mapped to another button.
For example, you can have North set to Autofire, and East set to normal Fire. That way, you can use both at the same time, for games such as R-Type which require holding down a button for a special move, but also would benefit from autofire at the same time.
This is not affected by the option of Autofire in the Controller/Input selection, which would still apply for all buttons if enabled.