Input: Get both Controller and Joystick names during joystick init
Fixes situation where a joystick is recognized as a controller, but it's being used in RetroArch environments (e.g. RetroPie).
Previously, the name used would be the controller name, but since RetroArch always uses the Joystick name instead, it would not recognize the mapping. This was the case with the USB Competition Pro on RetroPie.