enhancement: refactored input event handling to minimize latency
The previous code (inherited from WinUAE) was not very efficient, as it would constantly try to read from all supported controllers, and check the state of each supported button/axis per controller.

This commit refactors that logic to:
- Only read joystick or controller events when they are triggered
- Removed constant polling for each button/axis state
- Reduce CPU usage and minimize input latency

Furthermore, Joystick and Controller button events are handled separately, as a Controller button event will eventually also trigger a Joystick Button one. We only need to parse and act on one of them, however.