refactor: Fallback to A1200 config, if A600 ROM is not found (fixes #1183)
If the A600HD (2.05) Kickstart ROM is not available, Amiberry would try to use the A1200 3.0 one, but still using the CPU options from the XML. This of course failed, since that ROM requires an 020 CPU model.

Added a graceful fallback to A1200 config for such cases, where the CPU will also be set to 68020 in order to boot properly. Compatibility might be reduced, but at least it won't fail completely like before. The logs will also indicate that something like this has happened, to help troubleshoot things.