refactor: Don't populate "default" in config name automatically
The "default" config name is special for Amiberry, since if it exists, it will be loaded automatically on startup.
Previously, that would be populated in the Name field if no other config was loaded.
Changing this behavior to leave it empty, unless a config was loaded last (in which case, the field will contain the name of the last loaded config).