refactor: Only allow VKBD if screen is not RTG
The Virtual (On-Screen) keyboard was designed for native modes, mostly games which needed some keyboard input and the user might have only a controller connected, with no keyboard.
Under RTG environments, it doesn't make much sense to use an on-screen keyboard. The current implementation also is not dynamically scaling to higher resolutions, making it even more useful in an RTG environment.

Changing the behavior to only enable and toggle the VKBD in native screenmodes.