bugfix: WHDBooter would default to 020 CPU in cases it should use 68000
If the A600 ROM is not available, there is a failsafe to use an 68020 CPU and an A1200 config, which should cover most cases.
However, there was a scenario that was missed: if no CPU is specified in the XML, it's treaded as an "invalid CPU" case, which falls back to the 020 failsafe.
This resulted in cases where we had an A600 config, with an 020 CPU clocked at 7MHz, even though the A600 ROM was available in the system (instead of the correct and less demanding config of an 68000 CPU).

This fix adds an extra check to make sure that doesn't happen.