enhancement: remember window position and size (fixes #1049)
When starting emulation in Windowed mode, if the window is resized or moved, it would lose that information if we entered the GUI again and resumed emulation.
This changes that behavior, to re-use the previously stored X/Y and Width/Height values of the window, if they were set.

To reset those values to their defaults, one can enter the GUI and modify any of the Width/Height values or the AutoCrop setting.

# Conflicts:
#	src/osdep/gui/PanelDisplay.cpp