refactor: only enable Hardware RTG if using Virtual Mouse mode as well
Due to SDL2 limitations with the RelativeMouse mode (it hides the cursor), we cannot use a hardware cursor unless we also use absolute positioning (i.e. Virtual Mouse mode). Otherwise, we get a blank cursor, which is not very useful.
To avoid this situation, the option will now be disabled automatically, unless the Virtual Mouse mode is also enabled first.