enhancement: Added Save Mapping button in Custom Controls
This allows non-retroarch mappings to be saved to a controller file, similar to what Retroarch does.
In the current version, only the structure of the did->mapping is saved, which contains the hotkey and other custom buttons that can be set there.
If found, the file is loaded on startup as well. Retroarch mappings still take priority, this will only be used if we don't have retroarch at all, instead of the default mapping (which is still there, in case there's no controller mapping found)