bugfix: Don't force single_window mode in KMSDRM, fix crash when resuming from GUI
When single window mode was enabled, the emulation window was closed when the GUI one opened. However, if the multi-threaded drawing option was also enabled, the drawing thread would try to access the amiga surface, which was by now discarded, resulting in a segfault.

This changes the behavior as a workaround:
- don't force single window mode when KMSDRM is detected. Only certain window managers had a problem with that anyway, and it works in others.
- automatically disable the multi-threaded drawing option, if single window mode IS enabled.