bugfix: Improve CD32 default mapping (fixes #1237)
Amiberry would check if the button exists on the controller, before assigning a default mapping to it (like WinUAE does). In that check, it would also see if the button was reported as an axis, and in that case it would exclude it.
However, with some controllers that ended up having buttons not having a mapping on them, like the "Retroflag Classic USB gamepad", which reported RShoulder as such a case.

To fix this, I'm removing the check for the "isrealbutton" and instead provide the mapping directly. If the button doesn't exist, the event will never get triggered anyway, so this should be safe to do.