refactor: Create plugins directory, use it for external plugins (like floppybridge)
From now on, external libraries such as the CAPSimg.so and libfloppybridge.so, will live in a separate directory named "plugins". This cleans up the main dir from such files.

The new path is added to the GUI (Paths) and config files accordingly. The default value is <start_path>/plugins.

The Makefile will now build floppybridge by default - CAPSimg is a submodule, so it needs to be checked out before it can be built.

MacOS is a special case, since the libs there should be under the Resources folder. So, we're leaving that as-is for now.