refactor: Use CD32 model as a fallback for CD auto-detection (fixes #1325)
Instead of checking for CDTV, then CD32 and as a fallback going for A1200CD, we will look for CDTV and use CD32 as a fallback always.

Most of the titles will be in those two categories anyway, and many might not contain that in the file path, leading to incorrectly trying to boot them (and failing) as an A1200 CD config.