refactor: return to previous approach: GUI window is always Windowed, except on KMSDRM (#1351)
Due to how modern Window Managers handle things, if we open the GUI in full-window while the emulation screen is also opened in full-window, the second one gets minimized. That means that when we resume from the GUI, we don't get back to the emulation screen, even if we call for SDL_RaiseWindow() - it gets ignored.

To resolve this, I'm reverting this behavior to the previous approach: The GUI is always Windowed, and opens on top of the emulation screen. The exception is when we're using KMSDRM, which doesn't have the notion of windows.

But under KMSDRM we might run into a similar problem as well, so that needs a separate solution.