bugfix: Fix KMSDRM problems when switching between GUI -> emulation (fixes #1351)
The emulation screen would not get the focus always, when resuming from the GUI, if running under KMSDRM.
Removed the code that checked if the window had focus, and instead trying to grab it right away now.

It still sometimes doesn't get it (not sure why), but at least you can click on it once and resume properly now.