feat: Auto-detect MT32 titles in whdbooter
Auto-detect any titles supporting MT32 MIDI output in WHDBooter, and enable the relevant feature in Amiberry (if the ROMs are installed).
This currently works based on the filename (it has to contain the word "MT32"), as it's not included in the XML.
Not many title supported MT32, so maybe this is good enough.