enhancement: Added VSync options from WinUAE in Display panel (#1368)
Added the WinUAE VSync options for Native and RTG as dropdowns, instead of the single checkbox we had before.
The Lagless modes are not implemented, not even sure we can do that in SDL2.
The other modes work the same way as in WinUAE.

Amiberry still uses V-Sync internally for the SDL_Presenter, so this only changes what is passed internally in the emulation. These options only make sense to enable if the monitor is already set to the right refresh rate - i.e. don't enable if your monitor only does 60Hz and you want to emulate a PAL machine (which needs 50).