bugfix: remove newlines from debugger input
The "fd" command didn't work in the debugger. This was because the "d"
was followed in the input buffer by \n rather than \0.

The Windows version of console_get replaces the terminating \r with \0.
amiberry's version uses fgets, which doesn't do this; it did adjust the
length value returned, but the debugger code doesn't look at the length.

Make it actually strip the newline characters when adjusting the length.