enhancement: Added option to use JST instead of WHDLoad binary
In case someone needs this, there is now an option to use JST instead of WHDLoad for the Booter.

use_jst_instead_of_whd (boolean)

The option is saved in the global configuration of amiberry.conf and when enabled, will switch out the startup-sequence references from WHDLoad to JST, including any differences in options.

The old AMOS-based booter is not changed with this.