build: don't hardcode /usr prefix in the source
The new installation rules respect CMAKE_INSTALL_PREFIX and the other
GNUInstallDirs variables, but the prefix was hardcoded as /usr in
several places in osdep/amiberry.cpp, so it wouldn't actually work if
installed to a different prefix.

Define AMIBERRY_DATADIR and AMIBERRY_LIBDIR in CMakeLists.txt, and use
these instead of the hardcoded paths.