Added registry file, refactor things to use that for storing ROMs
Instead of storing the detected ROMs in the amiberry.conf file, we will switch to using the available functionality of registry -> ini file, that WinUAE already has.
This has become necessary, in order to use the same logic when storying the ROM entries (with their ID and type, not just the filename), which is used from the Expansions panel for custom board roms.

It also means changing the existing ROMs functionality to match, so it's likely some things will need fixing after this as well.