1545 savestates fixes (#1547)
* bugfix: set last active config to the savestate name (#1545)

* Enhance filename determination in gui_update()

The code now includes additional conditions to determine the `filename` in the `gui_update()` function. Specifically, it checks if `whdload_prefs.whdload_filename` is not empty and uses it to set the `filename`. If that condition is not met, it then checks if `last_loaded_config` has a non-zero length and uses it to set the `filename`. The previous logic that set `filename` to either `last_loaded_config` or "default.uae" has been removed.

* Improve savestate handling and UI feedback in PanelSavestate

Only enable buttons is a savestate title is set.
Show the title when displaying a not found message.
If a savestate is not found, don't try to load the screenshot for it either.
Minor code optimization.