enhancement: CDTV/CD32 auto-detect improvements (#1605)
When detecting CDTV or CD32 model from autoloading a CD image file, we might come across cases where the filename contains both of these strings. That would lead to CDTV being preferred always, since it was the first match.

Changing the logic to prefer CD32 in those cases, assuming that the newer model is a better choice (?)
Also, refactored the detection to only consider the filename, not the full path to it, to avoid directory names messing with the detection.