bugfix: the default_prefs used ECS chipset settings (#1607)
The default_prefs used the ECS chipset option, instead of OCS which is the default in Quickstart. This caused a difference in behavior, when loading something from the GUI vs when loading it from the command line, without opening the GUI. This is also apparent in WinUAE.

Changing this here to OCS, until we hear back from Toni, in case there is some reason for the difference.