refactor: use amiberry.conf for configured directories before creating the default ones (#1612, #1613)
Change the behavior on startup, regarding the default directories.
Check for the existence of the amiberry.conf file, and if found, load the settings from there, before attempting to create any missing directories.
This should fix the behavior where the default directories were always created, even if you had other paths configured in the conf file.

Also, add the config_path back to the amiberry.conf file, which was removed earlier.