Changes to have button-only gamepads completly mapped.
When a gamepad does provide only buttons, also for directions, SDL as well as
RetroArch expect the directions to be in the enum range from 11 to 14
(..._BUTTON_DPAD_..., starting with 0), thus expecting 15 buttons in miniumum.
In rare cases such gamepad does not provide 15 buttons physical buttons.  It
has the effect that the buttonn with indices [#phys_buttons,15) are not
recognized in an Amiberry game.  The Amiberry configuration GUI is not affected
as it uses SDL enum constants (e.g. SDL_CONTROLLER_BUTTON_DPAD_UP) to resolve
the pyhisical button from the map.  During the controller initialization this
map is populated with all 15 RetroArch buttons, regardless of the reported
physical button count.