fix all volatile variable breaking chained assignments (#1687)
warning: using value of assignment with ‘volatile’-qualified left
         operand is deprecated [-Wvolatile]

Chained assignments of volatile variables (var1 = var2 = var3 = 0;)
result in a read-modify-write opcode chain which breaks the idea behind
using volatile. Here during the modify phase the volatile variable can
be modified by an external source (the software running inside the
emulator), then the volatile variable is written again with the former
read value, completely dropping the value from the external source. It
is a rare scenario, but it happens, hence the warning from the compiler.
The fix is easy, break up the chained assignments.