refactor: fix build for RISC-V by making PCem optional (#1685)
RISC-V is not supported in PCem, so we'll need to make it optional, in order to allow the build to complete.
That also means disabling any dependency that uses PCem, as well:
- NCR
- NCR9X
- X86 boards
- QEMU CPU
- DRACO (uses X86 as well)

There is now a new option in the main CMakeLists.txt file, to control the PCem support. It defaults to ON, to keep the existing behavior. RISC-V platforms should have that option disabled.